refactor: category 도메인 dto 패키지 구분
