+++ src/main/java/org/ahpuh/surf/category/controller/CategoryController.java
@@ -1,10 +1,10 @@
 package org.ahpuh.surf.category.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.ahpuh.surf.category.dto.CategoryCreateRequestDto;
-import org.ahpuh.surf.category.dto.CategoryDetailResponseDto;
-import org.ahpuh.surf.category.dto.CategoryResponseDto;
-import org.ahpuh.surf.category.dto.CategoryUpdateRequestDto;
+import org.ahpuh.surf.category.dto.request.CategoryCreateRequestDto;
+import org.ahpuh.surf.category.dto.request.CategoryUpdateRequestDto;
+import org.ahpuh.surf.category.dto.response.CategoryDetailResponseDto;
+import org.ahpuh.surf.category.dto.response.CategoryResponseDto;
 import org.ahpuh.surf.category.service.CategoryService;
 import org.ahpuh.surf.jwt.JwtAuthentication;
 import org.springframework.http.ResponseEntity;
