+++ src/main/java/org/ahpuh/surf/category/converter/CategoryConverter.java
@@ -1,8 +1,8 @@
 package org.ahpuh.surf.category.converter;
 
-import org.ahpuh.surf.category.dto.CategoryCreateRequestDto;
-import org.ahpuh.surf.category.dto.CategoryDetailResponseDto;
-import org.ahpuh.surf.category.dto.CategoryResponseDto;
+import org.ahpuh.surf.category.dto.request.CategoryCreateRequestDto;
+import org.ahpuh.surf.category.dto.response.CategoryDetailResponseDto;
+import org.ahpuh.surf.category.dto.response.CategoryResponseDto;
 import org.ahpuh.surf.category.entity.Category;
 import org.ahpuh.surf.user.entity.User;
 import org.springframework.stereotype.Component;
